Refactor cache and cache_from_uri
