+++ xinference/model/utils.py
@@ -14,6 +14,7 @@
 import json
 import logging
 import os
+import shutil
 from json import JSONDecodeError
 from pathlib import Path
 from typing import Any, Callable, Dict, Optional, Tuple
@@ -21,6 +22,7 @@ from typing import Any, Callable, Dict, Optional, Tuple
 from fsspec import AbstractFileSystem
 
 from ..constants import XINFERENCE_CACHE_DIR, XINFERENCE_ENV_MODEL_SRC
+from .core import CacheableModelSpec
 
 logger = logging.getLogger(__name__)
 MAX_ATTEMPTS = 3
@@ -192,6 +194,133 @@ def is_valid_model_uri(model_uri: Optional[str]) -> bool:
     else:
         # TODO: handle other schemes.
         return True
+
+
+def cache_from_uri(
+    model_spec: CacheableModelSpec,
+    self_hosted_storage: bool = False,
+) -> str:
+    from fsspec import AbstractFileSystem, filesystem
+
+    cache_dir = os.path.realpath(
+        os.path.join(XINFERENCE_CACHE_DIR, model_spec.model_name)
+    )
+    if os.path.exists(cache_dir):
+        logger.info("cache %s exists", cache_dir)
+        return cache_dir
+
+    assert model_spec.model_uri is not None
+    src_scheme, src_root = parse_uri(model_spec.model_uri)
+    if src_root.endswith("/"):
+        # remove trailing path separator.
+        src_root = src_root[:-1]
+
+    if src_scheme == "file":
+        if not os.path.isabs(src_root):
+            raise ValueError(
+                f"Model URI cannot be a relative path: {model_spec.model_uri}"
+            )
+        os.makedirs(XINFERENCE_CACHE_DIR, exist_ok=True)
+        os.symlink(src_root, cache_dir, target_is_directory=True)
+        return cache_dir
+    elif src_scheme in ["s3"]:
+        # use anonymous connection for self-hosted storage.
+        src_fs: AbstractFileSystem = filesystem(src_scheme, anon=self_hosted_storage)
+        local_fs: AbstractFileSystem = filesystem("file")
+
+        files_to_download = []
+        os.makedirs(cache_dir, exist_ok=True)
+
+        for path, _, files in src_fs.walk(model_spec.model_uri):
+            for file in files:
+                src_path = f"{path}/{file}"
+                local_path = src_path.replace(src_root, cache_dir)
+                files_to_download.append((src_path, local_path))
+
+        from concurrent.futures import ThreadPoolExecutor
+
+        failed = False
+        with ThreadPoolExecutor(max_workers=min(len(files_to_download), 4)) as executor:
+            futures = [
+                (
+                    src_path,
+                    executor.submit(
+                        copy_from_src_to_dst, src_fs, src_path, local_fs, local_path
+                    ),
+                )
+                for src_path, local_path in files_to_download
+            ]
+            for src_path, future in futures:
+                if failed:
+                    future.cancel()
+                else:
+                    try:
+                        future.result()
+                    except:
+                        logger.error(f"Download {src_path} failed", exc_info=True)
+                        failed = True
+
+        if failed:
+            logger.warning(f"Removing cache directory: {cache_dir}")
+            shutil.rmtree(cache_dir, ignore_errors=True)
+            raise RuntimeError(
+                f"Failed to download embedding model '{model_spec.model_name}' "
+            )
+        return cache_dir
+    else:
+        raise ValueError(f"Unsupported URL scheme: {src_scheme}")
+
+
+def cache(model_spec: CacheableModelSpec, model_description_type: type):
+    if (
+        hasattr(model_spec, "model_uri")
+        and getattr(model_spec, "model_uri", None) is not None
+    ):
+        logger.info(f"Embedding model caching from URI: {model_spec.model_uri}")
+        return cache_from_uri(model_spec=model_spec)
+
+    cache_dir = os.path.realpath(
+        os.path.join(XINFERENCE_CACHE_DIR, model_spec.model_name)
+    )
+    if not os.path.exists(cache_dir):
+        os.makedirs(cache_dir, exist_ok=True)
+    meta_path = os.path.join(cache_dir, "__valid_download")
+    if valid_model_revision(meta_path, model_spec.model_revision):
+        return cache_dir
+
+    from_modelscope: bool = model_spec.model_hub == "modelscope"
+    if from_modelscope:
+        from modelscope.hub.snapshot_download import snapshot_download as ms_download
+
+        download_dir = retry_download(
+            ms_download,
+            model_spec.model_name,
+            None,
+            model_spec.model_id,
+            revision=model_spec.model_revision,
+        )
+        for subdir, dirs, files in os.walk(download_dir):
+            for file in files:
+                relpath = os.path.relpath(os.path.join(subdir, file), download_dir)
+                symlink_local_file(os.path.join(subdir, file), cache_dir, relpath)
+    else:
+        from huggingface_hub import snapshot_download as hf_download
+
+        retry_download(
+            hf_download,
+            model_spec.model_name,
+            None,
+            model_spec.model_id,
+            revision=model_spec.model_revision,
+            local_dir=cache_dir,
+            local_dir_use_symlinks=True,
+        )
+    with open(meta_path, "w") as f:
+        import json
+
+        desc = model_description_type(None, None, model_spec)
+        json.dump(desc.to_dict(), f)
+    return cache_dir
 
 
 def copy_from_src_to_dst(
